Huge refactor
Did a major refactor to the Engine class. It is now called the Renderer
and creates the framework for Vulkan use. Anything usage-specific, like
pipelines, descriptor sets, buffers, etc are now handled in
RenderSystems. All RenderSystems inherit from RenderSystem and must
implement a render() method. The Application then calls a method in the
Renderer class that calls each RenderSystem's render() method in the
order they were added.

I am also moving a lot of the setup code to the Application class. My
thinking is that the Application is any code which would not be a part
of an API. The rest of the code base is acting like an API. Working
through this has been tedious but I have been learning a lot about
structuring code from it. I just have to be careful not to get lost in
the trap of designing "perfectly organized" code.