Added camera class
The engine now has a camera class with orthographic or perspective
projections, as well as a few methods to set the view matrix.

Because of this, we now have an aspect-ratio-corrected circle!

Also added a global UBO to send view and projection matrices. I need to
sort through the existing descriptor sets and organize them better. I
also want to refactor the frame class and provide more frame data.