Added a timer class. Improved logger singleton.
Added a timer class that calculates frameTime and keeps track of FPS. It
is to be used for physics calculations involving time.

I also refactored
the logger after learning about a better way to implement singletons. It
makes more sense to me and is safer.

The particles now update their positions based off velocity and
bounce off the walls instead of teleporting on the opposite side.

This will make it easy to implement forces and accelerations.