BUGFIX: First few frames being very slow
In the Frame class constructor, the render fences were being initialized
unsignaled. This caused the vkWaitForFences function call to timeout for
the first frame of each swapchain image, aka the first three frames of
the engine. To fix, I just made the Frame constructor to initialize the
render fence to signaled.

I also made some adjustments to the particle system