Particles now collide
Added a very very naive collision detection algorithm (O(n^2)) with some
parameters such as damping factors, gravity, and # of simulation substeps.

Layed groundwork for mouse interaction with the simulation. I am
planning to add a push and pull relative to the mouse.

Working on an input event system to get input information to the
particle system.