Mouse interaction and input events
There is now mouse interaction when left or right clicking. Particles
will push or pull correspondingly.

In order to capture mouse inputs inside of the particle system, I set up
an event handling system inside the InputManager class. Now as long as a
system has access to the input manager, it can subscribe to input
events.