Ensure suitable physical device is a discrete GPU
Ran the code on a laptop and it was selecting the integrated graphics
device instead of the discrete one. Now it looks for a suitable device
and selects the default device if none are found.

I want to think of another solution to this later but for now this
works.