Shader class minor refactor
Shader class is no longer a namespace-type class with only static
methods. It now stores a shader module and stage flag. This is for ease
of deletion for the shader modules, since they don't need to be around
after creating a pipeline.

I also made all the pipeline builder methods return the pipeline builder
handle to be able to chain them together.

Next, I want to refactor the Pipeline class and change how pipeline
layouts are made