First pass at SPH-based fluid behavior
In its current state, it is not working properly. The particles are
supposed to come to an equilibrium state, but they just maintain their
level of high energy, potentially even increasing it over time. I am not
sure what the cause of this is. It's possible that everything is working
properly and I just have incorrect parameters set. Though I do get the
feeling that there is a bug.

I added a bunch of new functions to the ParticleSystem2D class to
calculate densities and pressure forces.

I also found and fixed a bug with minimizing the window since integrating ImGui.

I also made the world coordinate system larger and fixed an issue that
arose from that with the Hand class.