Start of proper SPH fluid simulation
I started writing the main chunk of the code for SPH fluid simulation.
This includes two different smoothing kernel functions, a density
calculation for each particle, and another acceleration source coming
from pressure. I had a lot of trouble getting the fluid to work
properly, but I think it is behaving as expected at the moment. I can
handle about 500 particles at about 30 fps, so the next step is
optimization using spatial hashing.

I tried turning on gravity, but I am not sure what is going on with
that, but I want to optimize the simulation first so i can put more
particles on the screen.