Implemented Compact Hashing
A spatial partitioning algorithm that allows the SPH computations to
skip all particles that are outside the smoothing radius. It does this
by dividing the space into a grid, giving each grid a hash value, and
associating particles in that grid cell by the hash value.

Went from being able to handle about 500 particles at 30 fps to over 2000