Undo previous change. Implemented differently
Instead of pausing rendering when the spacebar is pressed, just the
particle system updates are paused. This means you can still interact
with the gui if the system is paused. This is better and more flexible.
I have yet to implement the "do one frame" portion. I need to think
about how to do it better.