First implementation of RK4
The meat and potatoes of the solver code is there, but it is not working
properly yet. The code compiles, runs, and does not crash, but the
system completely breaks down after 3 frames due to incorrect
computations.

I am going to simplify the RK4 method into a modified euler to
try to narrow the issue down.