Shader overhaul
Translated shaders from GLSL to Slang

New class ShaderManager that deals with shader compilation. Eventually
it will deal with shader caching and reflection but that is TBD.

Vulkan shader modules are now created from either a filepath to a SPIR-V
compiled file, or a const uint32_t* pointer containing the code in
memory.

ShaderManager at the start now compiles all shaders at runtime and holds
them in memory. This will be changed since it currently holds all of the
shaders in memory, but I need to think about shader caching more carefully.